+++ 404.php
@@ -9,10 +9,11 @@
 get_header();
 ?>
 <section id="error-page" <?php echo prefix_template::AddContainer(prefix_template::$template_container, true); ?>>
-    <article>
-      <h1 class="post-title"><?php echo _e('404 Error', 'devTheme'); ?></h1>
-      <p><?php _e('Page could not be found.', 'devTheme'); ?></p>
-      <a href="<?php echo get_bloginfo('url'); ?>" class="button"><?php _e('Back to home', 'devTheme'); ?></a>
-    </article>
+  <?php echo prefix_template::get_thumbnail(); ?>
+  <article>
+    <h1 class="post-title"><?php echo _e('404 Error', 'devTheme'); ?></h1>
+    <p><?php _e('Page could not be found.', 'devTheme'); ?></p>
+    <a href="<?php echo get_bloginfo('url'); ?>" class="button"><?php _e('Back to home', 'devTheme'); ?></a>
+  </article>
 </section>
 <?php get_footer() ?>