+++ index.php
@@ -12,8 +12,18 @@ get_header();
 <section id="home" <?php echo prefix_template::AddContainer(prefix_template::$template_container, true); ?>>
   <div>
     <?php if (have_posts() ) : while (have_posts()) : the_post();
-      $blog_type = get_post_format(get_the_ID()) ? get_post_format(get_the_ID()) : "default";
-      get_template_part('template_parts/' . get_post_type() . '_' . $blog_type);
+      if(get_post_type() == "post" || post_type_supports(get_post_type(), 'post-formats')):
+        $blog_type = get_post_format(get_the_ID()) ? get_post_format(get_the_ID()) : "default";
+        // blog output
+        if(locate_template('template_parts/' . get_post_type() . '_' . $blog_type . '.php')):
+          get_template_part('template_parts/' . get_post_type() . '_' . $blog_type);
+        else:
+          get_template_part('template_parts/post_' . $blog_type);
+        endif;
+      else:
+        // default output
+        get_template_part('template_parts/post_default');
+      endif;
     endwhile; endif; ?>
   </div>
   <?php get_sidebar(); ?>