+++ template_parts/post_audio.php
@@ -24,28 +24,36 @@ endif;
 
 $css = 'temp-' . get_post_type() . '-image';
 $css .= $audio ? ' flex' : '';
+// disable content loading
+if($args['mediaOnly'] && $args['mediaOnly'] == 1):
+  $returnExcerpt = false;
+else:
+  $returnExcerpt = true;
+endif;
 ?>
 
 <article class="<?php echo $css; ?>">
   <?php
-    if($audio):
-      echo '<div>';
-    else:
-      echo '';
-    endif;
-  ?>
-
-    <?php get_template_part('template_parts/header'); ?>
-    <div class="entry-content">
-      <?php the_excerpt(); ?>
-    </div>
-    <?php get_template_part('template_parts/footer'); ?>
-
-    <?php
+    if($returnExcerpt):
       if($audio):
-        echo '</div>';
+        echo '<div>';
       else:
         echo '';
+      endif;
+  ?>
+
+      <?php get_template_part('template_parts/header'); ?>
+      <div class="entry-content">
+        <?php the_excerpt(); ?>
+      </div>
+      <?php get_template_part('template_parts/footer'); ?>
+
+    <?php
+        if($audio):
+          echo '</div>';
+        else:
+          echo '';
+        endif;
       endif;
     ?>